Initial commit with basic sync library.
This library is mostly convenience functions for now. It serves
to standardize the way sync is done among multiple Arduinos,
regardless of how the sync commands arrive.

If the sync is variable, this will not work. So be sure to resync
very close to when you need to be in sync.